Error pages fixed.

+++ project/src/main/java/com/catcare/project/Controller/Error/ClienteNotFoundException.java
@@ -1,17 +1,17 @@
 package com.catcare.project.Controller.Error;
 
 public class ClienteNotFoundException extends RuntimeException {
-    private int id;
+    private Long id;
 
-    public ClienteNotFoundException(int id) {
+    public ClienteNotFoundException(Long id) {
         this.id = id;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 }
